Refactor camera module to Compose structure
