+++ cameralibrary/src/main/java/com/cgfay/camera/activity/CameraComposeActivity.kt
@@ -1,10 +1,11 @@
-package com.cgfay.camera.compose
+package com.cgfay.camera.activity
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import com.cgfay.camera.compose.CameraPreviewViewModel
+import com.cgfay.camera.viewmodel.CameraPreviewViewModel
+import com.cgfay.camera.ui.CameraPreviewScreen
 
 class CameraComposeActivity : ComponentActivity() {
     private val previewViewModel: CameraPreviewViewModel by viewModels()
