+++ cameralibrary/src/main/java/com/cgfay/camera/ui/PreviewEffectScreen.kt
@@ -1,4 +1,4 @@
-package com.cgfay.camera.compose
+package com.cgfay.camera.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
