+++ cameralibrary/src/main/java/com/cgfay/camera/ui/PreviewResourceScreen.kt
@@ -1,4 +1,4 @@
-package com.cgfay.camera.compose
+package com.cgfay.camera.ui
 
 import android.net.Uri
 import androidx.compose.foundation.background
